implement lazy iterator API
Simplest way to create an iterator in Lua is defining a closure
that returns either an item or nil, where nil is the sentinel.

This change makes Surf:poly*() accept either tables or iterators.
Also, Frame:mapped() both accepts and returns iterators.

Lazily drawing a polygon from a shapefile now looks like this:
    geobbox, lens, geopgons = sf:read_polygons(index)
    surf:polygons(frame:mapped(geopgons), 1)