map: make caches smaller by using VLPs
VLP = variable-length pair

Each byte in a  VLP uses high nibble to store X  and low nibble to
store Y.  Signed is encoded as  unsigned by using odd  numbers for
positive and even numbers for negative.

Usually 95+% of points are compressed to 1 byte per point.

This is not as optimal as Rice Coding, but probably simpler, since
we don't have to deal with bit streams.